move test_prune
